Version 1.5.7
git-svn-id: https://plugins.svn.wordpress.org/fpw-category-thumbnails/trunk@644447 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ ajax/remove.php
@@ -1,24 +1,34 @@
-<?php
-		//	AJAX request to Remove thumbnails
-		$assignments = $this->getAssignmentsArray( $this->getAllCategories() );
-
-		while ( strlen( key( $assignments ) ) ) {
-			$catid = key( $assignments );
-			$parg = array(
-				numberofposts => -1,
-				nopaging => true,
-				category => $catid,
-				post_type => 'any' );
-			$posts = get_posts( $parg );
-			foreach ( $posts as $post ) {
-				$post_id = $post -> ID;
-				//	make sure this is not a revision
-				if ( 'revision' != $post -> post_type )
-					delete_post_meta( $post_id, '_thumbnail_id' );
-			}
-			next( $assignments );
-		}
-		reset( $assignments );
-		echo '<p><strong>' . __( 'All thumbnails removed successfully.', 'fpw-fct' ) . '</strong></p>';
-		die();
+<?php
+//	AJAX request to Remove thumbnails
+
+//	prevent direct access
+if ( preg_match( '#' . basename(__FILE__) . '#', $_SERVER[ 'PHP_SELF' ] ) )  
+	die( "Direct access to this script is forbidden!" );
+
+$assignments = $this->getAssignmentsArray( $this->getAllCategories() );
+
+while ( strlen( key( $assignments ) ) ) {
+	$catid = key( $assignments );
+
+	$parg = array(
+		'numberofposts' => -1,
+		'nopaging' => true,
+		'category' => $catid,
+		'post_type' => 'any' );
+
+	$posts = get_posts( $parg );
+
+	foreach ( $posts as $post ) {
+		$post_id = $post -> ID;
+		//	make sure this is not a revision
+		if ( 'revision' != $post -> post_type )
+			delete_post_meta( $post_id, '_thumbnail_id' );
+	}
+
+	next( $assignments );
+}
+
+reset( $assignments );
+echo '<p><strong>' . __( 'All thumbnails removed successfully.', 'fpw-fct' ) . '</strong></p>';
+die();
 ?>
